quarkus: consolidate knative integration tests
