+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-knative/src/main/resources/application.properties
@@ -21,8 +21,3 @@
 quarkus.log.console.enable = false
 quarkus.banner.enabled     = false
 
-#
-# Quarkus :: Camel
-#
-quarkus.camel.routes-discovery.enabled = false
-
