--- camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-knative/src/main/resources/env.json
@@ -1,13 +1,0 @@
-{
-  "services": [
-    {
-      "type": "endpoint",
-      "name": "from",
-      "metadata": {
-        "service.path": "/knative",
-        "camel.endpoint.kind": "source",
-        "knative.event.type": "camel.k.evt"
-      }
-    }
-  ]
-}
