+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-knative/src/test/resources/routes.properties
