--- camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-runtime-knative/src/test/resources/env.json
@@ -1,20 +1,0 @@
-{
-  "services": [
-    {
-      "type": "endpoint",
-      "name": "process",
-      "metadata": {
-        "camel.endpoint.kind": "source",
-        "service.path": "/knative"
-      }
-    },
-    {
-      "type": "endpoint",
-      "name": "process",
-      "metadata": {
-        "service.url": "http://localhost:{{env:HTTP_TEST_PORT}}/knative",
-        "camel.endpoint.kind": "sink"
-      }
-    }
-  ]
-}
